+++ README.md
@@ -1,14 +1,21 @@
 # Language Learning Chat
-AI chatbot web app that will carrying a conversation while providing feedback on your spelling and grammar.
+
+AI chatbot web app that will carry on a conversation while providing feedback on your spelling and
+grammar.
 
 ## Setup
+
 ### Frontend
+
 From the root of the directory:
+
 ```
 cd language-learning-chat-ui
 npm install
 npm start
 ```
-### LanguageTool Server
-Follow the docker setup instructions [here](https://hub.docker.com/r/silviof/docker-languagetool) to set up a local LanguageTool Server.
 
+### LanguageTool Server
+
+Follow the docker setup instructions [here](https://hub.docker.com/r/silviof/docker-languagetool) to
+set up a local LanguageTool Server.
