+++ api/app.js
@@ -1,0 +1,2 @@
+import express from 'express';
+import chatService from './routes/chatService';
