+++ ui/src/App.css
@@ -1,0 +1,12 @@
+.App-content {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  font-size: calc(10px + 2vmin);
+  background-color: white;
+}
+
+.App-link {
+  color: #61dafb;
+}
